+++ src/Force.php
@@ -17,14 +17,15 @@ use craft\services\Fields;
 use craft\services\Plugins;
 use craft\web\twig\variables\CraftVariable;
 use craft\web\UrlManager;
+use flipbox\craft\psr3\Logger;
 use flipbox\ember\helpers\UrlHelper;
+use flipbox\ember\modules\LoggerTrait;
 use flipbox\force\fields\Query as QueryField;
 use flipbox\force\fields\SObjects as SObjectIdsField;
 use flipbox\force\models\Settings as SettingsModel;
 use flipbox\force\patron\Events;
 use flipbox\force\web\twig\variables\Force as ForceVariable;
 use yii\base\Event;
-use yii\log\Logger;
 
 /**
  * @author Flipbox Factory <hello@flipboxfactory.com>
@@ -34,12 +35,36 @@ use yii\log\Logger;
  */
 class Force extends Plugin
 {
+    use LoggerTrait;
+
     /**
      * @inheritdoc
      */
     public function init()
     {
         parent::init();
+
+        // Components
+        $this->setComponents([
+            'cache' => services\Cache::class,
+            'connections' => services\Connections::class,
+            'elements' => services\Elements::class,
+            'logger' => [
+                'class' => Logger::class
+            ],
+            'queryField' => services\QueryField::class,
+            'queries' => services\Queries::class,
+            'resources' => services\Resources::class,
+            'sObjectAssociations' => services\SObjectAssociations::class,
+            'sObjectsField' => services\SObjectsField::class,
+            'transformers' => services\Transformers::class
+        ]);
+
+        // Modules
+        $this->setModules([
+            'cp' => cp\Cp::class
+
+        ]);
 
         // Fields
         Event::on(
@@ -88,6 +113,14 @@ class Force extends Plugin
                 }
             }
         );
+    }
+
+    /**
+     * @return string
+     */
+    protected static function getLogFileName(): string
+    {
+        return 'force';
     }
 
     /**
@@ -291,72 +324,5 @@ class Force extends Plugin
                 'force/settings/sobjects' => 'force/cp/settings/view/sobjects/index'
             ]
         );
-    }
-
-    /*******************************************
-     * LOGGING
-     *******************************************/
-
-    /**
-     * Logs a trace message.
-     * Trace messages are logged mainly for development purpose to see
-     * the execution work flow of some code.
-     * @param string $message the message to be logged.
-     * @param string $category the category of the message.
-     */
-    public static function trace($message, string $category = null)
-    {
-        Craft::getLogger()->log($message, Logger::LEVEL_TRACE, self::normalizeCategory($category));
-    }
-
-    /**
-     * Logs an error message.
-     * An error message is typically logged when an unrecoverable error occurs
-     * during the execution of an application.
-     * @param string $message the message to be logged.
-     * @param string $category the category of the message.
-     */
-    public static function error($message, string $category = null)
-    {
-        Craft::getLogger()->log($message, Logger::LEVEL_ERROR, self::normalizeCategory($category));
-    }
-
-    /**
-     * Logs a warning message.
-     * A warning message is typically logged when an error occurs while the execution
-     * can still continue.
-     * @param string $message the message to be logged.
-     * @param string $category the category of the message.
-     */
-    public static function warning($message, string $category = null)
-    {
-        Craft::getLogger()->log($message, Logger::LEVEL_WARNING, self::normalizeCategory($category));
-    }
-
-    /**
-     * Logs an informative message.
-     * An informative message is typically logged by an application to keep record of
-     * something important (e.g. an administrator logs in).
-     * @param string $message the message to be logged.
-     * @param string $category the category of the message.
-     */
-    public static function info($message, string $category = null)
-    {
-        Craft::getLogger()->log($message, Logger::LEVEL_INFO, self::normalizeCategory($category));
-    }
-
-    /**
-     * @param string|null $category
-     * @return string
-     */
-    private static function normalizeCategory(string $category = null)
-    {
-        $normalizedCategory = 'Force';
-
-        if ($category === null) {
-            return $normalizedCategory;
-        }
-
-        return $normalizedCategory . ': ' . $category;
     }
 }